requirements/default.txt
grpcio~=1.38
grpcio-tools~=1.38
protobuf~=3.17
PyQt5~=5.15
